content/media/test/test_readyState.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: readyState</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id='v1'></video><audio id='a1'></audio>
<pre id="test">
<script class="testbody" type="text/javascript">
"use strict";
var v1 = document.getElementById('v1');
var a1 = document.getElementById('a1');
var passed = "truthy";

is(v1.readyState, 0);
is(a1.readyState, 0);

try {
  v1.readyState = 0;
} catch (e) {
  passed = !passed;
}
try {
  a1.readyState = 0;
} catch (e) {
  passed = !passed;
}
ok(passed === true,
   "Setting readyState throws in strict mode (readonly attribute)");
</script>

<script class="testbody" type="text/javascript">
var v1 = document.getElementById('v1');
var a1 = document.getElementById('a1');
var passed = false;

is(v1.readyState, 0);
is(a1.readyState, 0);

try {
  v1.readyState = 1;
  a1.readyState = 1;
  passed = v1.readyState === 0 && a1.readyState === 0;
} catch(e) { }
ok(passed, "Should not be able to set readyState (readonly attribute)");
</script>
</pre>
</body>
</html>
